test case for testSentinelMasterListener
