+++ src/test/java/redis/clients/jedis/JedisSentinelTest.java
@@ -6,8 +6,10 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import org.junit.After;
 import org.junit.Before;
@@ -31,8 +33,12 @@ public class JedisSentinelTest {
   protected static HostAndPort sentinelForFailover = HostAndPorts.getSentinelServers().get(2);
   protected static HostAndPort masterForFailover = HostAndPorts.getRedisServers().get(5);
 
+  public final Set<String> sentinels = new HashSet<>();
+
   @Before
-  public void setup() throws InterruptedException {
+  public void setUp() throws Exception {
+    sentinels.clear();
+    sentinels.add(sentinelForFailover.toString());
   }
 
   @After
@@ -97,6 +103,35 @@ public class JedisSentinelTest {
     } finally {
       j.close();
       j2.close();
+    }
+  }
+
+  @Test
+  public void testSentinelMasterListener() throws InterruptedException {
+    Jedis j = new Jedis(sentinelForFailover);
+    Jedis j2 = new Jedis(sentinelForFailover);
+
+    SentinelPoolConfig config = new SentinelPoolConfig();
+    config.setEnableActiveDetectListener(true);
+    config.setEnableDefaultSubscribeListener(true);
+    config.setActiveDetectIntervalTimeMillis(5 * 1000);
+    config.setSubscribeRetryWaitTimeMillis(5 * 1000);
+
+    JedisSentinelPool pool = new JedisSentinelPool(FAILOVER_MASTER_NAME, sentinels, config, 1000,
+        "foobared", 2);
+
+    try {
+      HostAndPort masterGetFromPoolBefore = pool.getResource().connection.getHostAndPort();
+
+      JedisSentinelTestUtil.waitForNewPromotedMaster(FAILOVER_MASTER_NAME, j, j2);
+
+      HostAndPort masterGetFromPoolCurrent = pool.getResource().connection.getHostAndPort();
+
+      assertNotEquals(masterGetFromPoolBefore, masterGetFromPoolCurrent);
+    } finally {
+      j.close();
+      j2.close();
+      pool.destroy();
     }
   }
